+++ examples/resources/auth0_trigger_actions/import.sh
@@ -1,0 +1,4 @@
+# This resource can be imported using the bindings trigger ID.
+#
+# Example:
+terraform import auth0_trigger_actions.example "post-login"
